Combining nested loop MultiLambda tests.
